pyproject.toml
[tool.poetry]
name = "api-recommendation"
version = "0.1.0"
description = ""
authors = ["Ana Martínez Saucedo <anmartinez@uade.edu.ar>"]
readme = "README.md"
packages = [{include = "api_recommendation"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "2.25.1"
nltk = ">=3.7,<4.0"
gensim = ">=4.2.0,<4.3.0"
spacy = ">=3.3.1,<3.4.0"
ray = ">=2.1.0,<2.2.0"
numpy = ">=1.23.2,<1.24.0"
scikit-learn = ">=1.1.2,<1.2.0"
ruamel-yaml = "^0.18.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
